update file  reader

+++ Example/SJMediaCacheServer/SJTestViewController.m
@@ -31,7 +31,7 @@
     _startTime = MCSTimerStart();
     NSURLRequest *request = [NSURLRequest mcs_requestWithURL:[NSURL URLWithString:@"https://dh2.v.netease.com/2017/cg/fxtpty.mp4"] range:NSMakeRange(0, 10 * 1024 * 1024)];
     NSURLSessionDataTask *task = [_session dataTaskWithRequest:[request mcs_requestWithRedirectURL:[SJMediaCacheServer.shared playbackURLWithURL:request.URL]]];
-    NSLog(@"Task<%ld>.proxy resume", task.taskIdentifier);
+    NSLog(@"Task<%lu>.proxy resume", (unsigned long)task.taskIdentifier);
     [task resume];
 }
 
@@ -39,19 +39,19 @@
     _startTime = MCSTimerStart();
     NSURLRequest *request = [NSURLRequest mcs_requestWithURL:[NSURL URLWithString:@"https://dh2.v.netease.com/2017/cg/fxtpty.mp4"] range:NSMakeRange(0, 10 * 1024 * 1024)];
     NSURLSessionDataTask *task = [_session dataTaskWithRequest:request];
-    NSLog(@"Task<%ld>.network resume", task.taskIdentifier);
+    NSLog(@"Task<%lu>.network resume", (unsigned long)task.taskIdentifier);
     [task resume];
 }
 
 #pragma mark -
 
 - (void)URLSession:(NSURLSession *)session dataTask:(NSURLSessionDataTask *)dataTask didReceiveResponse:(NSURLResponse *)response completionHandler:(void (^)(NSURLSessionResponseDisposition))completionHandler {
-    NSLog(@"Task<%ld> response after (%lf) seconds.", dataTask.taskIdentifier, MCSTimerMilePost(_startTime));
+    NSLog(@"Task<%lu> response after (%lf) seconds.", (unsigned long)dataTask.taskIdentifier, MCSTimerMilePost(_startTime));
     completionHandler(NSURLSessionResponseAllow);
 }
 
 - (void)URLSession:(NSURLSession *)session task:(NSURLSessionTask *)task didCompleteWithError:(NSError *)error {
-    NSLog(@"Task<%ld> complete after (%lf) seconds.", task.taskIdentifier, MCSTimerMilePost(_startTime));
+    NSLog(@"Task<%lu> complete after (%lf) seconds.", (unsigned long)task.taskIdentifier, MCSTimerMilePost(_startTime));
     printf("\n");
 }
